+++ gazon-service/src/main/resources/db/changelog/v-1.3/02-fill-personal_address-table.sql
@@ -1,6 +1,0 @@
-INSERT INTO personal_address (address, directions, apartment, floor, entrance, door_code, post_code)
-VALUES ('Malaya Bronnaya St, Moscow, Russia', 'First entrance to the right after the alley', '404', '13', '1', '1234', '136881');
-INSERT INTO personal_address (address, apartment, floor, entrance)
-VALUES ('Malyy Cherkasskiy Pereulok, Moscow, Russia', '123', '6', '2');
-INSERT INTO personal_address (address, apartment, floor, entrance)
-VALUES ('Yuzhnaya St, Voronezh, Russia', '63', '1', '1');