+++ gazon-service/src/main/resources/db/changelog/v-1.3/04-fill-postomat-table.sql
@@ -1,6 +1,0 @@
-INSERT INTO postomat (address, directions, shelf_life_days)
-VALUES ('Gagarinskiy Pereulok, 65, Moscow, Russia', 'Inside the supermarket', 15);
-INSERT INTO postomat (address, shelf_life_days)
-VALUES ('Ulitsa Surikova, 31, Yekaterinburg, Sverdlovsk Oblast, Russia', 1);
-INSERT INTO postomat (address, shelf_life_days)
-VALUES ('Ligovsky Ave, 76, St Petersburg, Russia', 30);